#63 - refactoring: moved config and tests
