+++ src/test/java/de/bayern/gdi/config/ApplicationSettingsTest.java
@@ -15,9 +15,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.utils;
+package de.bayern.gdi.config;
 
-import de.bayern.gdi.config.ApplicationSettings;
 import junit.framework.TestCase;
 import org.junit.Test;
