+++ src/test/java/de/bayern/gdi/config/DownloadConfigTest.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.utils;
+package de.bayern.gdi.config;
 
 import junit.framework.TestCase;
 import org.junit.Test;
