+++ src/test/java/de/bayern/gdi/config/ServiceSettingsTest.java
@@ -15,9 +15,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.bayern.gdi.utils;
+package de.bayern.gdi.config;
 
-import de.bayern.gdi.config.ServiceSettings;
+import de.bayern.gdi.utils.XML;
 import junit.framework.TestCase;
 import org.junit.Test;
 import org.xml.sax.SAXException;
